requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.1.0
marker-pdf>=0.3.0
requests>=2.31.0
tenacity>=8.2.0
pytest>=7.4.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
slowapi>=0.1.9
